refactor(frontend): use providers index import
