+++ frontend/__tests__/.server/providers/container-config.provider.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
-import { ContainerConfigProviderImpl } from '~/.server/providers/container-config.provider';
+import { ContainerConfigProviderImpl } from '~/.server/providers';
 
 describe('ContainerConfigProviderImpl', () => {
   it("should call the container's get function with the correct identifier", () => {
