+++ frontend/__tests__/.server/providers/container-service.provider.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { mock } from 'vitest-mock-extended';
 
 import { SERVICE_IDENTIFIER } from '~/.server/constants';
-import { ContainerServiceProviderImpl } from '~/.server/providers/container-service.provider';
+import { ContainerServiceProviderImpl } from '~/.server/providers';
 
 describe('ContainerServiceProviderImpl', () => {
   it("should call the container's get function with the correct identifier", () => {
